feat: move get-started home section to shared components
