+++ public/animations/shared/get-started.riv
