+++ src/app/(home-page)/devs/page.jsx
@@ -1,6 +1,5 @@
 import AiIndex from 'components/pages/home/ai-index';
 import Bento from 'components/pages/home/bento';
-import GetStarted from 'components/pages/home/get-started';
 import Hero from 'components/pages/home/hero/hero';
 import Industry from 'components/pages/home/industry';
 import InstantProvisioning from 'components/pages/home/instant-provisioning';
@@ -8,6 +7,7 @@ import Lightning from 'components/pages/home/lightning';
 import Logos from 'components/pages/home/logos';
 import Multitenancy from 'components/pages/home/multitenancy';
 import Trusted from 'components/pages/home/trusted';
+import GetStarted from 'components/shared/get-started';
 import SEO_DATA from 'constants/seo-data';
 import getMetadata from 'utils/get-metadata';
 
@@ -27,7 +27,7 @@ const CodingForEntrepreuneursPage = () => (
     <Multitenancy />
     <Industry />
     <Trusted />
-    <GetStarted />
+    <GetStarted title="Features of tomorrow.<br /> Available today." />
   </>
 );
