+++ src/app/(home-page)/home/page.jsx
@@ -3,7 +3,6 @@ import { redirect } from 'next/navigation';
 import { checkCookie } from 'app/actions';
 import AiIndex from 'components/pages/home/ai-index';
 import Bento from 'components/pages/home/bento';
-import GetStarted from 'components/pages/home/get-started';
 import Hero from 'components/pages/home/hero/hero';
 import Industry from 'components/pages/home/industry';
 import InstantProvisioning from 'components/pages/home/instant-provisioning';
@@ -11,6 +10,7 @@ import Lightning from 'components/pages/home/lightning';
 import Logos from 'components/pages/home/logos';
 import Multitenancy from 'components/pages/home/multitenancy';
 import Trusted from 'components/pages/home/trusted';
+import GetStarted from 'components/shared/get-started';
 import SEO_DATA from 'constants/seo-data';
 import getMetadata from 'utils/get-metadata';
 
@@ -36,7 +36,7 @@ const HomePage = async () => {
       <Multitenancy />
       <Industry />
       <Trusted />
-      <GetStarted />
+      <GetStarted title="Features of tomorrow.<br /> Available today." />
     </>
   );
 };
